propogations_exporter: Add code to produce metrics on serial propogation
This script will listen to the jorunal log for a specific pattern for a
zone resign.

The script will parse out the new serial and then start to monitor all
downstream nameservers to get an idea of when the new serial porpogated
to the down streams.

We also add a metric to show if the serial is in sync accross all
downstreams as well as one to replicate rr count.

This will allows us to set up alerts of a zone remains out of sync for
an extended period of time or if the RR count changes by a large
percentage.